Modularize patching steps
